+++ src/main/java/com/iKeeper/homepage/domain/calendar/entity/Calendar.java
@@ -1,6 +1,6 @@
 package com.iKeeper.homepage.domain.calendar.entity;
 import com.iKeeper.homepage.domain.calendar.dto.request.CalendarRequest;
-import com.iKeeper.homepage.global.entity.Field;
+import com.iKeeper.homepage.domain.user.entity.Field;
 import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.DynamicInsert;
