+++ src/main/java/com/iKeeper/homepage/domain/post/dao/mapping/PostList.java
@@ -1,10 +1,5 @@
 package com.iKeeper.homepage.domain.post.dao.mapping;
 
-import com.iKeeper.homepage.domain.user.entity.Grade;
-import com.iKeeper.homepage.domain.user.entity.Major;
-import com.iKeeper.homepage.domain.user.entity.Status;
-import com.iKeeper.homepage.global.entity.Field;
-
 import java.time.LocalDateTime;
 
 public interface PostList {
