+++ src/main/java/com/iKeeper/homepage/domain/user/entity/UserRole.java
@@ -1,0 +1,5 @@
+package com.iKeeper.homepage.domain.user.entity;
+
+public enum UserRole {
+    GUEST, USER, ADMIN
+}
