--- src/main/java/com/iKeeper/homepage/global/entity/UserRole.java
@@ -1,5 +1,0 @@
-package com.iKeeper.homepage.global.entity;
-
-public enum UserRole {
-    GUEST, USER, ADMIN
-}
